+++ src/models/FrequencyRange.ts
@@ -1,0 +1,6 @@
+interface FrequencyRange {
+  lower: number
+  upper: number
+}
+
+export { FrequencyRange };
